KnightPi_Mission_Commander_RPi_v1_2/requirements.txt
fastapi
uvicorn[standard]
httpx
pyserial
pydantic
pydantic-settings
aiofiles
